ticker-style location row in event cards
The location row was inconsistent across views: in the Week stream's
EventCard it shared the time row and truncated with "…" for long names;
in the Day tab's DayEventRow it had no lineLimit and wrapped to multiple
lines, breaking the row's vertical rhythm.

Move MarqueeText into its own Components file so it can be reused, and
make it collapse to zero height for empty text (so the Day masthead
doesn't reserve a blank line when there's no city). Apply it to:

- DayEventRow: location now sits on a single marqueeing row beneath
  the title; rows without a location stay at one line.
- EventCard timed style: split the inline time + location into two
  rows — time on its own row, then the location row only when present.
